add concourse-web terraform module
concourse-web is the big module which does a lot of things, because you
can't really run a concourse without the web piece
this module does:

- load balancer provisioning
- key generation
- dns entries and cert provisioning
- autoscaling group / launch config for web
- iam roles
- database provisioning

the concourse web created is highly available

load balancer provisioning
--------------------------

this provisions 2 load balancers: an ALB and an ELB

The ALB is external and used for TLS termination for public traffic and
has an external DNS entry of $deployment.gds-reliability.engineering

The ELB is internal and is used for communication between the worker
nodes and the web nodes.

key generation
--------------

In order to have a working concourse we need some rsa keys to be
generated, these are stored using ssm parameter store and encrypted
using kms.

The keys are as follows:

(KMS) worker shared key, for decrypting secrets, this is available to
all workers as a encryption/decryption key, they can't read the
parameters used by other workers. The web node can access it as well so
it can read/write secrets .

(KMS) web key, for decrypting secrets, this is available to all web
nodes. This is used for web specific secrets (i.e. the RSA keys in
parameter store).

(RSA) web ssh key, this is the ssh key used by web nodes and trusted by
the worker nodes, the private key is available to web, and the public
key in openssh format is available to the workers

(RSA) web session key, this is used for session signing and is only
available to the web nodes

(RSA) worker ssh keys, these keys are different to each worker pool and
used for communicating with the web nodes via ssh. All the public keys are
available to the web nodes but not the private keys.

DNS and ACM
-----------

We need an internal DNS zone so the web nodes can find the database and
the worker nodes can find the web load balancer.

We need external DNS records for provisioning an ACM cert for TLS, as
well as the DNS record for the ALB.

ASG and launch config
---------------------

We need some nodes to run the concourse scheduler and web ui. This is
just cloud init ubuntu 18.04 which has SSM on it by default. The nodes
startup and download concourse. Web nodes run prometheus-node-exporter
and the concourse web nodes also emit metrics which can be consumed by
prometheus.

IAM
---

The web nodes have an instance profile which allow them to access
secrets for the web instance as well as secrets for pipelines. The web
instance also needs to access parameter store in order to read secrets.

Database provisioning
---------------------

Concourse needs a database with which the web nodes communicate. This is
provisioned in this module, it is encrypted at rest and in transit. It
is private.

Network connectivity
--------------------

We don't have an egress proxy, the web nodes can talk to the worker
nodes, the worker nodes can talk to the web nodes via a load balancer
for healthchecking. The ALB and ELB can talk to the web nodes.

Network connectivity is done on a point/point basis using security
groups so we don't really care what ports are open/locked down, we
shouldn't care if the ALB is using 8080 or 8081, likewise if the ELB is
using 2222 or 2223, the important bit is which component has which
security group.