add concourse-worker-pool module
This module is for creating worker pools which are the underlying EC2
instances which run jobs on behalf of team.

In the subnets module there is the network isolation, in this module
there is the IAM isolation.

A concourse team in our deployments has a 1-1 mapping with an IAM
instance profile (i.e. a role). This means that we can give different
tenants different permissions, and tenants can allow their concourse to
assume roles in their AWS accounts without other tenants being able to
do so.

This module provisions an ASG where the nodes run the concourse worker
and the prometheus-node-exporter. On startup the worker registers itself
with the web nodes.

The worker node is configured to be ephemeral so we do not have to
manually prune workers