Use worker role arns directly in concourse-keys
Before the worker roles were created in separate modules, we've
refactored this behaviour to be done in the concourse-keys module.

If we use the role arns directly it introduces an explicit dependency in
the terraform graph which forces the roles to be created before the kms
key policy is updated, this fixes the race condition when provisioning a
new team.

When applying this change it is a no-op

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>