Store local user password in SSM
The initial need for the story is to allow self-updating pipelines from
inside the shared concourse.

In order for these users to be useful, we need to expose the password to
the pipeline so it could use it where appropriate.

Concourse is configured to read secrets from the SSM under certain path
and allow reading them from inside team pipelines.