[cyber/security-audit-role] bump to terraform 0.12
This upgrades the module to terraform 0.12.

Terraform 0.12 has been out for a while.  Continuing to use tf 0.11
syntax in here causes errors in consumers, such as:

```
Error: Unsupported block type

  on .terraform/modules/cyber_security_audit_role/cyber-security/modules/gds_security_audit_role/gds_security_audit_role.tf line 4, in data "template_file" "trust":
   4:   vars {

Blocks of type "vars" are not expected here. Did you mean to define argument
"vars"? If so, use the equals sign to assign it a value.
```

This commit upgrades to tf 0.12, mostly by running the handy
`terraform 0.12upgrade` command.

Consumers should pin to a commit hash so this shouldn't break existing
users.