Revert one of the idiomatic v0.12 Terraform changes
I don't particularly want to spend any time figuring out why this breaks but
without reverting this change we see errors like this using Terraform v0.12.20:

```
Error: Invalid index

  on .terraform/modules/concourse_pay_subnets/reliability-engineering/terraform/modules/concourse-subnets/routes.tf line 52, in resource "aws_route_table_association" "concourse_public":
  52:   route_table_id = aws_route_table.concourse_public[count.index].id
    |----------------
    | aws_route_table.concourse_public is object with 6 attributes
    | count.index is 1

The given key does not identify an element in this collection value.
```

What's curious is that this seems to work fine for the equivalent
`aws_route_table_association` for the private subnet.