set logGroupName to splunk service instance guid
We need a unique value in the CSLS log events to use on the CSLS
processor side of the kinesis stream for mapping which log events end up
in which index in splunk.

The CSLS processor[1] is responsible for mapping log events to indexes
and it currently performs this task by looking up a mapping of
LogGroupName to index from a TOML configuration[2]

The Adapter was previously setting the LogGroupName to a static string,
so the log events did not have enough information to be mapped in this
manor.

This changes the log events output by the Adapter to use the
cloudfoundry service instance GUID as the log group name, so that the
GUID can be used to map logs sent to "service instances" to "splunk
indexes".

To do this the Broker has been modified to include the cloudfoundry
service instance GUID in the signed drain URL and then setting the
LogGroupName to the instance GUID (after verifying the URL has not been
tampered with).

The service instance GUID was chosen (rather than something more human
readable like say the Org Name or Space Name) for the following reasons:

* Neither the Broker or Adapter have direct access to the Org/Space name
  and would require communication with the cloudfoundry API to fetch the
  data. Which would have significantly increased the complexity of the
  applications and required maintaining broadly scoped credentials
* Org and Space names are mutable. If a tenant renamed their org, they
  would likely not expect this to break their log shipping.
* The service instance GUID is unique to the service, so with this
  method it is possible to create multiple services within the same
  space and have logs shipped to different indexes.
* The space GUID was another good option, (it's unique, immutable and
  available in the data provided to the Broker) but the ability to share
  service instances between spaces blurs the lines of which space the
  logs are really coming from.

[1] https://github.com/alphagov/centralised-security-logging-service/tree/master/kinesis_processor
[2] https://github.com/alphagov/centralised-security-logging-service/blob/master/kinesis_processor/accounts_loggroup_index.toml