use propper UUID types
string representations of GUIDs/UUIDs can be formatted in many ways, we
don't know if the platform will always provide the IDs in the exact same
format. If the platform decided to change it's GUID representation then
the following could occur:

* Logs sent to the Adapter would start failing verification as we relied
  on the string representation to calculate the HMAC
* The log-group name used in the outbound log events could change as it
  was based on the verbatim string representation of the service
  instance id from the platform

To avoid these issues, and protect further against invalid requests, we
parse the GUIDs into their underlaying bytes so that which ever string
representation is used the end result will be the same.

To make sure that we have a consitant string format for guid used in the
output log event log group name additional tests are added to confirm
output is always a lowercased form.