add a pipeline to apply cd/cd-staging terraform
extracts the main team password from existing terraform state, generates
a provider override config (so that the terraform can continue to use
target-based config when doing it by hand).

applies terraform against the environment.

to roll the web nodes a job scales the autoscaling group up to it's
max size, waits for it to settle, then attempts to land the old workers

once landed, another job scales back in (by running the terraform again)
causing the autoscaling group (which is now set with a termination
policy of "kill the oldest first") to remove old nodes.

The series of events is performed to give concourse the best chance of
being able to do a clean run without hitting "worker disapeared" due to
having the node pulled out from under a running task, however it's
unlikely to be perfect.