tag successful cd deploy pipeline runs
we need to be able to mark commits that have passed through the
deployment pipeline (in staging) so they can be picked up and deployed
(in prod).

the "test" in this case is simply using some resources that people
expect to work like being able to use the s3 resource, build a docker
container and push to ecr image repository.

we push a tag back to each repo using a task that force pushes a task
ref... the git-resource is not
used as it's tagging implimentation conflicts with it's "paths"
implementation. git-resource would only be useful if we were triggering
on ALL commits to master, but we are not as it is a shared monorepo and
we don't want to re-deploy concourse on every commit.