automate deploying concourse from upstream release
our current process for upgrading concourse is bumping version numbers
and hash in the terraform module, this isn't exactly difficult work but
it is toil and we are regually a couple of minor versions behind.

by monitoring the github releases in the pipeline, and automatically
deploying straight to staging we can automate this process a fair bit.

We can use the concourse UI to "pin" the resource to a known good
version in the production deployment and only move it on when we are
happy that it was ok in staging.