[concourse-worker] Watch for spot interruption notices
This adds a new service which polls instance metadata every 5 seconds
for [spot interruption notices][1].  If an interruption notice is
found, it sends a SIGUSR2 to concourse-worker, which [retires the worker][2].

Also, use `systemctl enable --now` to start services to cut a few
lines from the script.

Note that currently no workers actually run on spot instances; this
just makes it plausible that they might in future.

[1]: https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/spot-interruptions.html
[2]: https://concourse-ci.org/concourse-worker.html#gracefully-removing-a-worker