gracefully shut workers down
This configures the concourse-worker systemd unit to use SIGUSR2 to
stop workers.  SIGUSR2 is the "retire worker" signal - it causes the
worker to stop accepting new jobs, but finish any jobs it currently
has.

When an instance is terminated by EC2, for example during instance
rolling, systemd is used to perform a graceful shutdown of the system.
This means that, during node rolling, we will retire workers
gracefully rather than having workers just disappear while running
jobs.

This also updates the check-spot-interruption job to do a simple
`systemctl stop` instead of trying to do clever things with `systemctl
kill`.

Supersedes #122.