concourse: consistent test retries
we expect the tests to run into intermitant concourse errors immediately
after deploy while the main team's workers are settling.

we addressed this initially with some "attempts: 50", to retry on
errors, but didn't add these attempts to the rest as the job evolved.

50 was overkill ... 10 is plenty (in cases where errors are hit, it
usually recovers on like try 2 or 3).

add attempts to all the test steps