Add pipeline build-gdsre-docker
The intention for this pipeline is to build `gdsre` docker hub
organisation images on concourse rather than ad hoc on people's
laptops.

Currently it pushes to ECR rather than docker hub as a proof of
concept, but if people think it's a sensible thing to do, we can a)
change it to push to docker hub (involves adding docker hub creds to
the pipeline), b) add all the other `gdsre` dockerfiles too.