concourse: fix bad gen roll instanaces
if a worker is missing/dead when the deploy step runs then no "roll
instances" pipeline will be created for it.

this fetches the list of teams and merges it with the list of workers to
ensure that any known teams show up with at least one worker.

we also move this step to after scale-in, since teams will not be
updated until concourse-web has come up.