AUT-825: Create a new service for Big Grafana
The current service for Big Grafana has a service discovery but its details are empty. According to AWS documentation, updating existing services to configure service discovery for the first time or change the current configuration is not supported. This change deletes the current service and create a new service called <deployment>-concourse-grafana-v2 with service discovery enabled.

Author: @adityapahuja