extract new module `concourse-teams`
The idea here is that everything that's associated with a team should
live here.  We should aim to migrate things like the per-team S3
bucket and ECR repository out of concourse-worker-pool and into
concourse-teams; but that would involve some terraform state hackery
so we should be careful with that step.