block access to the AWS metadata service for global workers
If a worker can run jobs on behalf on any team, access to its IAM role
should be forbidden.  This is because accessing this role grants
access to the worker SSH key, which allows the bearer to impersonate a
worker, register with the web node, and get given secrets intended for
other teams.