concourse: increase idle timeout
we are seeing 504 responses when fetching builds.

increases the timeout at the ALB is not a "fix" for this, but it should
at least let us get the responses (and I believe get them temporarily
cached). We can then look into what is causing these requests to be so
slow.