disable concourse build retention options
We're seeing significant performance degradation after upgrading to
6.6.0, including massive increase in DB read throughput (up by a
factor of 1000-2000!).

We have a hunch that this might be caused by
https://github.com/concourse/concourse/pull/6051 and that
https://github.com/concourse/concourse/pull/6125 (not yet released)
might fix it.  Until then, our theory is that disabling these build
retention options will make concourse happier for now.

Plan is to release this and observe the results.  If things improve,
make a note to restore these when we upgrade to the next concourse
version.

Co-authored-by: Alex Monk <alex.monk@digital.cabinet-office.gov.uk>