[concourse] only specify major engine_version
We enable automatic minor engine version changes[1] which means that
engine_version can change outside the lifecycle of terraform.

It's fiddly to keep it all in sync because upgrades happen in a
staggered fashion.  For example, right now cd-staging-concourse is
10.13 but cd-concourse is 10.10; for grafana it's the other way round.

The aws provider docs[2] suggest we can specify an engine prefix to
make terraform ignore minor version changes.

[1]: https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/USER_UpgradeDBInstance.Upgrading.html#USER_UpgradeDBInstance.Upgrading.AutoMinorVersionUpgrades
[2]: https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/db_instance#engine_version