images: add build for govsvc/aws-terraform
the aws-terraform image is an image containing "terraform" and "aws-cli"
used in many concourse tasks.

previously aws-terraform images were hand tweaked/built each time
a terraform version was required and pushed to gdsre/aws-terraform

we now create a build step in the build images pipeline that tracks
terraform versions and automates new builds of the aws-terraform image
and pushes it to the govsvc/aws-terraform repository (for consistancy
with the others)

we don't trigger on new ubuntu versions, but pull whatever digest is
latest at the time a new terraform image comes out, this avoids
overwriting old tags. So you are expected to upgrade to newer terraform
versions to get the newer images. If a specific ubuntu image is
required in the future, it could be pinned in the resource by setting
the tag.