Reboot machines on init
To ensure they're running up to date software, in the case where it
was updated as part of the init process.

Rather than having services start, then stop because of the reboot,
this also changes the init process to just enable services, so that
they'll be first started after the reboot.