[cd] Deploy concourse using instance refresh
Currently, the deployment pipeline manages to use concourse to deploy
concourse by scaling itself out and in again.  This could be made
simpler by using the new [Instance Refresh feature][1].  This means
that we don't have to change the desired number of workers, or re-run
terraform, or anything.

Note: we still have the problem that, while polling the API to see if
all workers have been refreshed, the working doing the polling might
get terminated.  This commit doesn't (yet) fix that, we'll need to
address this.

Note2: there's probably a bunch of files in
reliability-engineering/pipelines/tasks/ that can be deleted now, I
haven't got round to it though.

[1]: https://aws.amazon.com/blogs/compute/introducing-instance-refresh-for-ec2-auto-scaling/