concourse-monitoring: make terraform 0.13 compliant
the main thing here is the need to move the grafana provider out of the
module, which requires the admin password be set/generated in the root
module too. therefore we take this as a new var to be passed in.

though the terraform 0.13 upgrade guide seems quite insistent that provider
requirements belong in the root module, we appear to need to provide the
`versions.tf` file, here autogenerated by the `terraform 0.13upgrade`
command, to map the namespaced providers to their local in-module names
otherwise terraform is unhappy.

and because the grafana provider's url is hard coded (elsewhere) and no
longer restricted by the dependency graph, we add some manual depends_on
settings to those resources that use the grafana provider so as there
isn't an attempt to put them in place before the grafana instances are
actually up.