[cd] ensure cloudwatch event rules are namespaced
I hit a problem where `main` team credentials weren't being rotated
correctly in prod.  This turned out to be because the `cd-staging`
deployment had clobbered the cloudwatch event rule that triggers
credential refresh for `main` and `sandbox` teams.  (Other teams don't
exist in `cd-staging` so were not affected.)

This adds `var.deployment` to the event rule name so that it doesn't
clobber things.