[cd] set MinHealthyPercentage=100
Instance Refresh has a habit of terminating our only `main` worker.
This breaks everything, because then web can't schedule any tasks.

The default MinHealthyPercentage is 90.  It's documented
[here](https://docs.aws.amazon.com/autoscaling/ec2/APIReference/API_RefreshPreferences.html)
as:

> The amount of capacity in the Auto Scaling group that must remain healthy during an instance refresh to allow the operation to continue, as a percentage of the desired capacity of the Auto Scaling group (rounded up to the nearest integer). The default is 90.

Given the "rounded up to the nearest integer" bit, I'm not hopeful
that this will fix things for us, but we should try it anyway.