Create automate/task-toolbox docker image
This is effectively a fork of GSP's concourse-task-toolbox that can continue
to be used in Verify and Big Concourse. It doesn't have things like kubectl.
It is only pushed to GHCR, since we wouldn't want to replace the GSP
task-toolbox as some stuff still uses it, and our docker registry of choice is
now GHCR rather than DockerHub.