cd roll instances pipeline: ensure asg is rolled in before starting procedure
an asg which for some reason is not at its minimum instance count
(likely due to scheduled scaling) when this procedure is followed would
not end up with entirely new workers. though it would likely exist
for a period where it had zero workers (because we land them *all*,
assuming new workers already exist), which is also not good.

so first run a "scale in", which in most cases should be a no-op. in
cases where it *does* shut down workers, those workers will be shut
down less gracefully than the process normally would because we don't
pre-land them - we can't because we don't know the names of the workers
which the asg will choose to shut down.

ideally we will make workers self-drain gracefully at some point, making
all of this work much nicer.